+++ app/src/main/java/nextstep/signup/ui/signin/EmailTextField.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import nextstep.signup.R
-import nextstep.signup.ui.SignupValidator.ResultType
+import nextstep.signup.ui.signin.SignupValidator.ResultType
 import nextstep.signup.ui.component.SignupTextField
 
 @Composable
